240423(화) JWT 검증함수 외부 모듈화 완료 &  JWT 토큰 기능 적용 (2) & 로그인 여부에 따른  좋아요 확인 구현 & 도서 조회 pagination 기능 구현(1)

+++ controller/LikeController.js
@@ -1,25 +1,34 @@
+const verifyJWT = require('../verifyJWT');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
 const jwt = require('jsonwebtoken');
-const dotenv = require('dotenv');
-dotenv.config();
 
 const addLike = (req,res)=>{
     const book_id = req.params.id;
 
     let decodedJwt = verifyJWT(req);
 
-    let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
-    let values = [decodedJwt.id, book_id];
-    conn.query(sql, values, 
-        (err, results)=>{
-            if(err){
-                console.log(err)
-                return res.status(StatusCodes.BAD_REQUEST).end();
+    if(decodedJwt instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message : "로그인 세션이 만료되었습니다. 다시 로그인해 주십시오"
+        });
+    }else if(decodedJwt instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message : "잘못된 JWT 토큰입니다."
+        });
+    }else{
+        let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
+        let values = [decodedJwt.id, book_id];
+        conn.query(sql, values, 
+            (err, results)=>{
+                if(err){
+                    console.log(err)
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.CREATED).json(results);
             }
-            return res.status(StatusCodes.CREATED).json(results);
-        }
-    ); 
+        ); 
+    }
 }
 
 const deleteLike = (req,res)=>{
@@ -27,24 +36,27 @@ const deleteLike = (req,res)=>{
 
     let decodedJwt = verifyJWT(req);
 
-    let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
-    let values = [decodedJwt.id, book_id];
-    conn.query(sql, values, 
-        (err, results)=>{
-            if(err){
-                console.log(err)
-                return res.status(StatusCodes.BAD_REQUEST).end();
+    if(decodedJwt instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message : "로그인 세션이 만료되었습니다. 다시 로그인해 주십시오"
+        });
+    }else if(decodedJwt instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message : "잘못된 JWT 토큰입니다."
+        });
+    }else{
+        let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
+        let values = [decodedJwt.id, book_id];
+        conn.query(sql, values, 
+            (err, results)=>{
+                if(err){
+                    console.log(err)
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.CREATED).json(results);
             }
-            return res.status(StatusCodes.CREATED).json(results);
-        }
-    ); 
-}
-
-function verifyJWT(req){
-    const receivedJwt = req.headers["authorization"];
-    const decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-    
-    return decodedJwt
+        ); 
+    }
 }
 
 module.exports = {
